Moved EnumType to appropriate package and added apache license and comments.
